restructured the program
Put the routines that calculate the equilibrium quantities (e.q. metric factors,
Z, R, ...) into a new module  eq_ops. This is called every time the calculations
are done  for a  certain alpha. The  Richardson loop is  supposed to  happen for
every alpha, after the equilibrium.
Also, the  order of  the matrices  g_C, h_C, g_V  and h_V  is changed  to employ
better the array storage system in Fortran.